ispravke potrebne za endpointove

+++ src/main/java/org/raflab/studsluzba/services/GrupaService.java
@@ -9,7 +9,11 @@ import org.raflab.studsluzba.repositories.GrupaRepository;
 import org.raflab.studsluzba.repositories.SkolskaGodinaRepository;
 import org.raflab.studsluzba.repositories.StudijskiProgramRepository;
 import org.raflab.studsluzba.utils.Converters;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -38,7 +42,17 @@ public class GrupaService {
         return grupaRepository.findAll();
     }
 
+    @Transactional
     public void deleteById(Long id) {
-        grupaRepository.deleteById(id);
+        if (!grupaRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "Entitet sa ID " + id + " ne postoji.");
+        }
+        try {
+            grupaRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT,
+                    "Ne moze se obrisati entitet jer postoje povezani zapisi.");
+        }
     }
 }
