+++ src/main/java/org/raflab/studsluzba/services/IspitPrijavaService.java
@@ -8,7 +8,11 @@ import org.raflab.studsluzba.model.entities.StudentIndeks;
 import org.raflab.studsluzba.repositories.IspitPrijavaRepository;
 import org.raflab.studsluzba.repositories.IspitRepository;
 import org.raflab.studsluzba.repositories.StudentIndeksRepository;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -41,7 +45,17 @@ public class IspitPrijavaService {
         return (List<IspitPrijava>) ispitPrijavaRepository.findAll();
     }
 
+    @Transactional
     public void deleteById(Long id) {
-        ispitPrijavaRepository.deleteById(id);
+        if (!ispitPrijavaRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "Entitet sa ID " + id + " ne postoji.");
+        }
+        try {
+            ispitPrijavaRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT,
+                    "Ne moze se obrisati entitet jer postoje povezani zapisi.");
+        }
     }
 }
