+++ src/main/java/org/raflab/studsluzba/services/IspitniRokService.java
@@ -6,7 +6,11 @@ import org.raflab.studsluzba.model.entities.IspitniRok;
 import org.raflab.studsluzba.model.entities.SkolskaGodina;
 import org.raflab.studsluzba.repositories.IspitniRokRepository;
 import org.raflab.studsluzba.repositories.SkolskaGodinaRepository;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -38,7 +42,18 @@ public class IspitniRokService {
         return (List<IspitniRok>) ispitniRokRepository.findAll();
     }
 
+    @Transactional
     public void deleteById(Long id) {
-        ispitniRokRepository.deleteById(id);
+        if (!ispitniRokRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "Entitet sa ID " + id + " ne postoji.");
+        }
+        try {
+            ispitniRokRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT,
+                    "Ne moze se obrisati entitet jer postoje povezani zapisi.");
+        }
     }
+
 }
