+++ src/main/java/org/raflab/studsluzba/services/NastavnikZvanjeService.java
@@ -7,7 +7,11 @@ import org.raflab.studsluzba.model.entities.NastavnikZvanje;
 import org.raflab.studsluzba.repositories.NastavnikRepository;
 import org.raflab.studsluzba.repositories.NastavnikZvanjeRepository;
 import org.raflab.studsluzba.utils.Converters;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -34,7 +38,17 @@ public class NastavnikZvanjeService {
         return nastavnikZvanjeRepository.findAll();
     }
 
+    @Transactional
     public void deleteById(Long id) {
-        nastavnikZvanjeRepository.deleteById(id);
+        if (!nastavnikZvanjeRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                    "Entitet sa ID " + id + " ne postoji.");
+        }
+        try {
+            nastavnikZvanjeRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT,
+                    "Ne moze se obrisati entitet jer postoje povezani zapisi.");
+        }
     }
 }
